requirements.txt
skale.py==4.0dev12
Click==7.1.2
python-dotenv==0.10.3
texttable==1.6.2